Add SolrCommonFinna trait.
